Apply local patch which creates an 'Configure Konsole' menu and dialog.
Some profile options don't actually belong to profiles.  They should be
in the main Konsole settings.  This is the first step is trying to
seperate the options.  I only did 1 option for this patch.
At some time it may be possible to put the profile dialog inside this
main settings dialog.