Add a 'Pre Bookmark command option' to the new settings dialog.
Instead of using a Profile option in the new Settings dialog, a
pre-bookmark command is added.  This allows the user to select if they
want a Ctrl+C, Ctrl+C + clear or nothing to be done before issuing a
bookmark command.
This patch restores the 'save window size' option to the profile page.